Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -2,7 +2,7 @@
 int main(){
     float n1, n2;
     scanf("%f%f", &n1, &n2);
-    int prod = n1 * n2;
+    float prod = n1 * n2;
     printf("Product: %.2f", prod);
     return 0;
 }
